chore: move crdt bar chart components
