+++ src/components/pages/race/breakouts/bar-chart/bar-chart.module.scss
